requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=0.15.0=pypi_0
astunparse=1.6.3=pypi_0
ca-certificates=2023.08.22=h06a4308_0
cachetools=5.3.2=pypi_0
certifi=2023.11.17=pypi_0
charset-normalizer=3.3.2=pypi_0
contourpy=1.1.1=pypi_0
cycler=0.12.1=pypi_0
flatbuffers=1.12=pypi_0
fonttools=4.46.0=pypi_0
gast=0.3.3=pypi_0
google-auth=2.25.2=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.32.0=pypi_0
h5py=2.10.0=pypi_0
idna=3.6=pypi_0
importlib-metadata=7.0.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.4.5=pypi_0
libedit=3.1.20221030=h5eee18b_0
libffi=3.2.1=hf484d3e_1007
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
markdown=3.5.1=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.6.0=pypi_0
ncurses=6.4=h6a678d5_0
numpy=1.19.2=pypi_0
oauthlib=3.2.2=pypi_0
openssl=1.1.1w=h7f8727e_0
opt-einsum=3.3.0=pypi_0
packaging=23.2=pypi_0
pillow=10.1.0=pypi_0
pip=23.3.1=py38h06a4308_0
protobuf=3.20.3=pypi_0
pyasn1=0.5.1=pypi_0
pyasn1-modules=0.3.0=pypi_0
pyparsing=3.1.1=pypi_0
python=3.8.0=h0371630_2
python-dateutil=2.8.2=pypi_0
readline=7.0=h7b6447c_5
requests=2.31.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
scipy=1.8.0=pypi_0
setuptools=68.2.2=py38h06a4308_0
six=1.15.0=pypi_0
sqlite=3.33.0=h62c20be_0
tensorboard=2.11.2=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow=2.4.4=pypi_0
tensorflow-estimator=2.4.0=pypi_0
termcolor=1.1.0=pypi_0
tk=8.6.12=h1ccaba5_0
typing-extensions=3.7.4.3=pypi_0
urllib3=2.1.0=pypi_0
werkzeug=3.0.1=pypi_0
wheel=0.41.2=py38h06a4308_0
wrapt=1.12.1=pypi_0
xz=5.4.5=h5eee18b_0
zipp=3.17.0=pypi_0
zlib=1.2.13=h5eee18b_0
